explicitely changed types

+++ src/AB04MD.f
@@ -28,12 +28,12 @@ C
 C     P       (input) INTEGER
 C             The number of system outputs.  P >= 0.
 C
-C     ALPHA,  (input) DOUBLE PRECISION
+C     ALPHA,  (input) REAL*16
 C     BETA    Parameters specifying the bilinear transformation.
 C             Recommended values for stable systems: ALPHA = 1,
 C             BETA = 1.  ALPHA <> 0, BETA <> 0.
 C
-C     A       (input/output) DOUBLE PRECISION array, dimension (LDA,N)
+C     A       (input/output) REAL*16 array, dimension (LDA,N)
 C             On entry, the leading N-by-N part of this array must
 C             contain the state matrix A of the original system.
 C             On exit, the leading N-by-N part of this array contains
@@ -43,7 +43,7 @@ C
 C     LDA     INTEGER
 C             The leading dimension of array A.  LDA >= MAX(1,N).
 C
-C     B       (input/output) DOUBLE PRECISION array, dimension (LDB,M)
+C     B       (input/output) REAL*16 array, dimension (LDB,M)
 C             On entry, the leading N-by-M part of this array must
 C             contain the input matrix B of the original system.
 C             On exit, the leading N-by-M part of this array contains
@@ -53,7 +53,7 @@ C
 C     LDB     INTEGER
 C             The leading dimension of array B.  LDB >= MAX(1,N).
 C
-C     C       (input/output) DOUBLE PRECISION array, dimension (LDC,N)
+C     C       (input/output) REAL*16 array, dimension (LDC,N)
 C             On entry, the leading P-by-N part of this array must
 C             contain the output matrix C of the original system.
 C             On exit, the leading P-by-N part of this array contains
@@ -63,7 +63,7 @@ C
 C     LDC     INTEGER
 C             The leading dimension of array C.  LDC >= MAX(1,P).
 C
-C     D       (input/output) DOUBLE PRECISION array, dimension (LDD,M)
+C     D       (input/output) REAL*16 array, dimension (LDD,M)
 C             On entry, the leading P-by-M part of this array must
 C             contain the input/output matrix D for the original system.
 C             On exit, the leading P-by-M part of this array contains
@@ -77,7 +77,7 @@ C     Workspace
 C
 C     IWORK   INTEGER array, dimension (N)
 C
-C     DWORK   DOUBLE PRECISION array, dimension (LDWORK)
+C     DWORK   REAL*16 array, dimension (LDWORK)
 C             On exit, if INFO = 0, DWORK(1) returns the optimal value
 C             of LDWORK.
 C
@@ -170,19 +170,19 @@ C
 C     ******************************************************************
 C
 C     .. Parameters ..
-      DOUBLE PRECISION  ZERO, ONE, TWO
+      REAL*16  ZERO, ONE, TWO
       PARAMETER         ( ZERO=0.0D0, ONE=1.0D0, TWO=2.0D0 )
 C     .. Scalar Arguments ..
       CHARACTER         TYPE
       INTEGER           INFO, LDA, LDB, LDC, LDD, LDWORK, M, N, P
-      DOUBLE PRECISION  ALPHA, BETA
+      REAL*16  ALPHA, BETA
 C     .. Array Arguments ..
       INTEGER           IWORK(*)
-      DOUBLE PRECISION  A(LDA,*), B(LDB,*), C(LDC,*), D(LDD,*), DWORK(*)
+      REAL*16  A(LDA,*), B(LDB,*), C(LDC,*), D(LDD,*), DWORK(*)
 C     .. Local Scalars ..
       LOGICAL           LTYPE
       INTEGER           I, IP
-      DOUBLE PRECISION  AB2, PALPHA, PBETA, SQRAB2
+      REAL*16  AB2, PALPHA, PBETA, SQRAB2
 C     .. External Functions ..
       LOGICAL           LSAME
       EXTERNAL          LSAME
