+++ src/AB07MD.f
@@ -26,7 +26,7 @@ C
 C     P       (input) INTEGER
 C             The number of system outputs.  P >= 0.
 C
-C     A       (input/output) DOUBLE PRECISION array, dimension (LDA,N)
+C     A       (input/output) REAL*16 array, dimension (LDA,N)
 C             On entry, the leading N-by-N part of this array must
 C             contain the original state dynamics matrix A.
 C             On exit, the leading N-by-N part of this array contains
@@ -35,7 +35,7 @@ C
 C     LDA     INTEGER
 C             The leading dimension of array A.  LDA >= MAX(1,N).
 C
-C     B       (input/output) DOUBLE PRECISION array, dimension
+C     B       (input/output) REAL*16 array, dimension
 C             (LDB,MAX(M,P))
 C             On entry, the leading N-by-M part of this array must
 C             contain the original input/state matrix B.
@@ -45,7 +45,7 @@ C
 C     LDB     INTEGER
 C             The leading dimension of array B.  LDB >= MAX(1,N).
 C
-C     C       (input/output) DOUBLE PRECISION array, dimension (LDC,N)
+C     C       (input/output) REAL*16 array, dimension (LDC,N)
 C             On entry, the leading P-by-N part of this array must
 C             contain the original state/output matrix C.
 C             On exit, the leading M-by-N part of this array contains
@@ -56,7 +56,7 @@ C             The leading dimension of array C.
 C             LDC >= MAX(1,M,P) if N > 0.
 C             LDC >= 1 if N = 0.
 C
-C     D       (input/output) DOUBLE PRECISION array, dimension
+C     D       (input/output) REAL*16 array, dimension
 C             (LDD,MAX(M,P))
 C             On entry, if JOBD = 'D', the leading P-by-M part of this
 C             array must contain the original direct transmission
@@ -110,7 +110,7 @@ C     .. Scalar Arguments ..
       CHARACTER         JOBD
       INTEGER           INFO, LDA, LDB, LDC, LDD, M, N, P
 C     .. Array Arguments ..
-      DOUBLE PRECISION  A(LDA,*), B(LDB,*), C(LDC,*), D(LDD,*)
+      REAL*16  A(LDA,*), B(LDB,*), C(LDC,*), D(LDD,*)
 C     .. Local Scalars ..
       LOGICAL           LJOBD
       INTEGER           J, MINMP, MPLIM
