+++ src/AB09JV.f
@@ -75,7 +75,7 @@ C             The dimension of the output vector, if JOB = 'V', or
 C             of the input vector, if JOB = 'C', of the system with
 C             the transfer-function matrix V.  PV >= 0.
 C
-C     A       (input) DOUBLE PRECISION array, dimension (LDA,N)
+C     A       (input) REAL*16 array, dimension (LDA,N)
 C             The leading N-by-N part of this array must contain the
 C             state matrix A of the system with the transfer-function
 C             matrix G in a real Schur form.
@@ -83,7 +83,7 @@ C
 C     LDA     INTEGER
 C             The leading dimension of the array A.  LDA >= MAX(1,N).
 C
-C     B       (input) DOUBLE PRECISION array, dimension (LDB,M)
+C     B       (input) REAL*16 array, dimension (LDB,M)
 C             The leading N-by-M part of this array must contain
 C             the input/state matrix B of the system with the
 C             transfer-function matrix G. The matrix BS is equal to B.
@@ -91,7 +91,7 @@ C
 C     LDB     INTEGER
 C             The leading dimension of the array B.  LDB >= MAX(1,N).
 C
-C     C       (input/output) DOUBLE PRECISION array, dimension (LDC,N)
+C     C       (input/output) REAL*16 array, dimension (LDC,N)
 C             On entry, the leading P-by-N part of this array must
 C             contain the output matrix C of the system with the
 C             transfer-function matrix G.
@@ -102,7 +102,7 @@ C
 C     LDC     INTEGER
 C             The leading dimension of the array C.  LDC >= MAX(1,P,PV).
 C
-C     D       (input/output) DOUBLE PRECISION array, dimension (LDD,M)
+C     D       (input/output) REAL*16 array, dimension (LDD,M)
 C             On entry, the leading P-by-M part of this array must
 C             contain the feedthrough matrix D of the system with the
 C             transfer-function matrix G.
@@ -114,7 +114,7 @@ C
 C     LDD     INTEGER
 C             The leading dimension of the array D.  LDD >= MAX(1,P,PV).
 C
-C     AV      (input/output) DOUBLE PRECISION array, dimension (LDAV,NV)
+C     AV      (input/output) REAL*16 array, dimension (LDAV,NV)
 C             On entry, the leading NV-by-NV part of this array must
 C             contain the state matrix AV of the system with the
 C             transfer-function matrix V.
@@ -134,7 +134,7 @@ C
 C     LDAV    INTEGER
 C             The leading dimension of the array AV.  LDAV >= MAX(1,NV).
 C
-C     EV      (input/output) DOUBLE PRECISION array, dimension (LDEV,NV)
+C     EV      (input/output) REAL*16 array, dimension (LDEV,NV)
 C             On entry, if JOBEV = 'G', the leading NV-by-NV part of
 C             this array must contain the descriptor matrix EV of the
 C             system with the transfer-function matrix V.
@@ -157,7 +157,7 @@ C             The leading dimension of the array EV.
 C             LDEV >= MAX(1,NV), if JOBEV = 'G';
 C             LDEV >= 1,         if JOBEV = 'I'.
 C
-C     BV      (input/output) DOUBLE PRECISION array,
+C     BV      (input/output) REAL*16 array,
 C             dimension (LDBV,MBV), where MBV = P, if JOB = 'V', and
 C             MBV = PV, if JOB = 'C'.
 C             On entry, the leading NV-by-MBV part of this array must
@@ -172,7 +172,7 @@ C
 C     LDBV    INTEGER
 C             The leading dimension of the array BV.  LDBV >= MAX(1,NV).
 C
-C     CV      (input/output) DOUBLE PRECISION array, dimension (LDCV,NV)
+C     CV      (input/output) REAL*16 array, dimension (LDCV,NV)
 C             On entry, the leading PCV-by-NV part of this array must
 C             contain the output matrix CV of the system with the
 C             transfer-function matrix V, where PCV = PV, if JOB = 'V',
@@ -188,7 +188,7 @@ C             The leading dimension of the array CV.
 C             LDCV >= MAX(1,PV) if JOB = 'V';
 C             LDCV >= MAX(1,P)  if JOB = 'C'.
 C
-C     DV      (input) DOUBLE PRECISION array,
+C     DV      (input) REAL*16 array,
 C             dimension (LDDV,MBV), where MBV = P, if JOB = 'V', and
 C             MBV = PV, if JOB = 'C'.
 C             The leading PCV-by-MBV part of this array must contain
@@ -207,7 +207,7 @@ C     IWORK   INTEGER array, dimension (LIWORK)
 C             LIWORK =   0,    if JOBEV = 'I';
 C             LIWORK = NV+N+6, if JOBEV = 'G'.
 C
-C     DWORK   DOUBLE PRECISION array, dimension (LDWORK)
+C     DWORK   REAL*16 array, dimension (LDWORK)
 C             On exit, if INFO = 0, DWORK(1) returns the optimal value
 C             of LDWORK.
 C
@@ -322,7 +322,7 @@ C
 C     ******************************************************************
 C
 C     .. Parameters ..
-      DOUBLE PRECISION  ZERO, ONE
+      REAL*16  ZERO, ONE
       PARAMETER         ( ZERO = 0.0D0, ONE = 1.0D0 )
 C     .. Scalar Arguments ..
       CHARACTER         DICO, JOB, JOBEV, STBCHK
@@ -330,20 +330,20 @@ C     .. Scalar Arguments ..
      $                  LDD, LDDV, LDEV, LDWORK, M, N, NV, P, PV
 C     .. Array Arguments ..
       INTEGER           IWORK(*)
-      DOUBLE PRECISION  A(LDA,*), AV(LDAV,*), B(LDB,*), BV(LDBV,*),
+      REAL*16  A(LDA,*), AV(LDAV,*), B(LDB,*), BV(LDBV,*),
      $                  C(LDC,*), CV(LDCV,*), D(LDD,*), DV(LDDV,*),
      $                  DWORK(*), EV(LDEV,*)
 C     .. Local Scalars ..
       CHARACTER*1       EVTYPE, STDOM
       LOGICAL           CONJS, DISCR, STABCK, UNITEV
-      DOUBLE PRECISION  ALPHA, DIF, SCALE, TOLINF, WORK
+      REAL*16  ALPHA, DIF, SCALE, TOLINF, WORK
       INTEGER           I, IA, IERR, KAI, KAR, KB, KC, KE, KF, KQ, KW,
      $                  KZ, LDW, LDWN, LW, SDIM
 C     .. Local Arrays ..
       LOGICAL           BWORK(1)
 C     .. External Functions ..
       LOGICAL           DELCTG, LSAME
-      DOUBLE PRECISION  DLAMCH, DLANGE
+      REAL*16  DLAMCH, DLANGE
       EXTERNAL          DELCTG, DLAMCH, DLANGE, LSAME
 C     .. External Subroutines ..
       EXTERNAL          AB09JX, DGEMM, DGGES, DLACPY, DLASET, DSWAP,
